Build score builder

+++ components/History/index.tsx
@@ -19,9 +19,15 @@ const index = ({ patternId }) => {
   })
 
   const handleDelete = (recordId) => {
-    deleteRecord({ variables: { recordId } })
-      .then(() => toast.success('Deleted!'))
-      .catch((e) => toast.error(e.message))
+    if (
+      window.confirm(
+        'Are you sure you want to delete this record? This action is irreversable.'
+      )
+    ) {
+      deleteRecord({ variables: { recordId } })
+        .then(() => toast.success('Deleted!'))
+        .catch((e) => toast.error(e.message))
+    }
   }
 
   if (!data || isEmpty(data.records)) return null
